+++ src/main/java/com/toty/user/application/UserService.java
@@ -2,7 +2,6 @@ package com.toty.user.application;
 
 import com.toty.Tag;
 import com.toty.following.domain.FollowingRepository;
-import com.toty.user.domain.LoginProvider;
 import com.toty.user.domain.User;
 import com.toty.user.domain.UserLink;
 import com.toty.user.domain.UserLinkRepository;
@@ -11,14 +10,12 @@ import com.toty.user.domain.UserTag;
 import com.toty.user.domain.UserTagRepository;
 import com.toty.user.presentation.dto.LinkDto;
 import com.toty.user.presentation.dto.request.UserInfoUpdateRequest;
-import com.toty.user.presentation.dto.request.UserSignUpRequest;
 import com.toty.user.presentation.dto.response.UserInfoResponse;
 import jakarta.transaction.Transactional;
 import java.io.File;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -35,22 +32,6 @@ public class UserService {
     @Value("${user.img-path}")
     private String basePath;
 
-    public Long signUp(UserSignUpRequest userSignUpRequest) {
-        if(userRepository.findByEmail(userSignUpRequest.getEmail()).isPresent()) {
-            throw new IllegalArgumentException("이미 가입된 사용자입니다.");
-        }
-        String hashedPwd = BCrypt.hashpw(userSignUpRequest.getPassword(), BCrypt.gensalt());
-
-        User user = User.builder()
-                .email(userSignUpRequest.getEmail())
-                .password(hashedPwd)
-                .nickname(userSignUpRequest.getNickname())
-                .phoneNumber(userSignUpRequest.getPhoneNumber())
-                .loginProvider(LoginProvider.FORM)
-                .build();
-        return userRepository.save(user).getId();
-    }
-
     // 본인 확인
     public boolean isSelfAccount(User user, Long id){
         if (id == user.getId()) {
@@ -58,7 +39,6 @@ public class UserService {
         }
         return false;
     }
-
 
     public UserInfoResponse getUserInfoResponse(User user, Long id) {
         UserInfoResponse userInfo;
@@ -125,24 +105,7 @@ public class UserService {
             // todo 예외 시
             // throw new ExpectedException(ErrorCode.FileIOException);
         }
-
-
     }
-
-    public String validateEmail(String email) {
-        if (userRepository.existsByEmail(email)) {
-            throw new IllegalArgumentException("사용할 수 없는 이메일입니다.");
-        }
-        return email;
-    }
-
-    public String validateNickname(String nickname) {
-        if (userRepository.existsByNickname(nickname)) {
-            throw new IllegalArgumentException("사용할 수 없는 이메일입니다.");
-        }
-        return nickname;
-    }
-
 
     public void deleteUser(User user, Long id) { // soft delete
         User foundUser = userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
