+++ src/main/java/com/toty/user/presentation/UserApiController.java
@@ -4,15 +4,13 @@ import com.toty.annotation.CurrentUser;
 import com.toty.following.application.FollowService;
 import com.toty.following.presentation.dto.response.FollowingListResponse;
 import com.toty.user.application.UserService;
+import com.toty.user.application.UserSignUpService;
 import com.toty.user.domain.User;
-import com.toty.user.domain.UserRepository;
 import com.toty.user.presentation.dto.request.UserInfoUpdateRequest;
 import com.toty.user.presentation.dto.request.UserSignUpRequest;
 import com.toty.user.presentation.dto.response.UserInfoResponse;
 import jakarta.servlet.http.HttpSession;
 import jakarta.websocket.server.PathParam;
-import java.io.File;
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +18,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -32,14 +29,14 @@ import org.springframework.web.multipart.MultipartFile;
 public class UserApiController {
 
     private final UserService userService;
+    private final UserSignUpService userSignUpService;
     private final FollowService followService;
 
     // 회원 가입
     @PostMapping("/")
     @ResponseBody
     public ResponseEntity<Long> signUp(@RequestBody UserSignUpRequest userSignUpRequest) {
-
-        Long userId = userService.signUp(userSignUpRequest);
+        Long userId = userSignUpService.signUp(userSignUpRequest);
         return new ResponseEntity<>(userId, HttpStatus.CREATED);
     }
 
@@ -66,7 +63,7 @@ public class UserApiController {
     @GetMapping("/check-email") // 파라미터로 받아오기
     @ResponseBody
     public ResponseEntity emailValidation(@RequestParam(name = "email") String email) {
-        String response = userService.validateEmail(email);
+        String response = userSignUpService.validateEmail(email);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
@@ -74,7 +71,7 @@ public class UserApiController {
     @GetMapping("/check-nickname")
     @ResponseBody
     public ResponseEntity nicknameValidation(@RequestParam(name = "nickname") String nickname) {
-        String response = userService.validateNickname(nickname);
+        String response = userSignUpService.validateNickname(nickname);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
